Move DB_SAFE_MINIMUM, DB_SAFE_MAXIMUM constants and db_safe_priority function into BaseOperator
Move DB_SAFE_MINIMUM, DB_SAFE_MAXIMUM constants and db_safe_priority function into BaseOperator
