+++ task-sdk/src/airflow/sdk/__init__.py
@@ -41,7 +41,6 @@ __all__ = [
     "PokeReturnValue",
     "TaskGroup",
     "Variable",
-    "WeightRule",
     "XComArg",
     "asset",
     "chain",
@@ -104,7 +103,6 @@ __lazy_imports: dict[str, str] = {
     "SecretCache": ".execution_time.cache",
     "TaskGroup": ".definitions.taskgroup",
     "Variable": ".definitions.variable",
-    "WeightRule": ".definitions.weight_rule",
     "XComArg": ".definitions.xcom_arg",
     "asset": ".definitions.asset.decorators",
     "chain": ".bases.operator",
