+++ task-sdk/src/airflow/sdk/__init__.pyi
@@ -51,7 +51,6 @@ from airflow.sdk.definitions.param import Param as Param
 from airflow.sdk.definitions.taskgroup import TaskGroup as TaskGroup
 from airflow.sdk.definitions.template import literal as literal
 from airflow.sdk.definitions.variable import Variable as Variable
-from airflow.sdk.definitions.weight_rule import WeightRule as WeightRule
 from airflow.sdk.definitions.xcom_arg import XComArg as XComArg
 from airflow.sdk.execution_time.cache import SecretCache as SecretCache
 from airflow.sdk.io.path import ObjectStoragePath as ObjectStoragePath
@@ -80,7 +79,6 @@ __all__ = [
     "SecretCache",
     "TaskGroup",
     "Variable",
-    "WeightRule",
     "XComArg",
     "asset",
     "chain",
