+++ task-sdk/src/airflow/sdk/definitions/_internal/abstractoperator.py
@@ -36,8 +36,8 @@ from airflow.sdk.definitions._internal.node import DAGNode
 from airflow.sdk.definitions._internal.setup_teardown import SetupTeardownContext
 from airflow.sdk.definitions._internal.templater import Templater
 from airflow.sdk.definitions.context import Context
+from airflow.sdk.definitions.weight_rule import WeightRule
 from airflow.utils.trigger_rule import TriggerRule
-from airflow.utils.weight_rule import WeightRule
 
 if TYPE_CHECKING:
     import jinja2
